BUGFIX: logINFO and render
the first logINFO bug wat not even implemented in ui.lua than a bug, the second bug was in render function as the function requires only a one argument, the second argument had to be only for C++ for the both SDLandlibusb mode. It was not supposed to be in lua whatsoever.